Changes for porting to IRIX64.  Important change is the addition of _MIPS_SZLONG which is how IRIX64 defines it has a 64 bit long. Fixed a bug in cdf.c.  When it tried to figure out the numrecs of a variable, it assumed the incoming pointer was for a 4 bytes long.  It was replaced with a temperary int32 variable as a go between. xdrposix.c has important changes.  netCDF uses XDR which stores "long" as 32 bits.  This causes confusion if the incoming integer is actually a long integer which is 64 bits.  IRIX64 provides two routine to store a 32-bit integer value, one for long arguments, one for int argument.  My change assumes netCDF routines are invoked with a 32-bit, not a long.  It has to go one way or the other and I had to make a call. There are still errors in the Fortran netCDF interface.

+++ mfhdf/xdr/xdrtest.c
@@ -70,7 +70,7 @@ char *av[] ;
 #endif
 
 	/* some random numbers, divisible by 4 and less than 32k */
-	static unsigned long seeks[] = 
+	static u_int seeks[] = 
 	{
 		   16828 ,
 		    8448 ,
@@ -91,7 +91,7 @@ char *av[] ;
 		   12556 ,
 	} ;
 
-	unsigned long poses[9] ;
+	u_int poses[9] ;
 	int jj = 0 ;
 
 	static char text[32] ={ "Hiya sailor. New in town?" };
